auto_batcher's jar file file invoker to right folder
